+++ test/DependencyInjection/Fixtures/php/ignore_errors_integration_overridden.php
@@ -8,7 +8,6 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 $container->loadFromExtension('sentry', [
     'options' => [
         'integrations' => [
-            'App\\Sentry\\Integration\\FooIntegration',
             'Sentry\\Integration\\IgnoreErrorsIntegration',
         ],
     ],